+++ stream-compositions/services/product-composition-service/src/main/java/com/backbase/stream/compositions/product/core/config/ProductConfigurationProperties.java
@@ -1,11 +1,9 @@
 package com.backbase.stream.compositions.product.core.config;
 
-import com.backbase.stream.product.task.BatchProductGroupTask;
+import com.backbase.stream.compositions.product.core.model.RequestConfig;
 import com.backbase.stream.product.task.BatchProductIngestionMode;
-import lombok.Data;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
+import lombok.experimental.SuperBuilder;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 import java.util.ArrayList;
@@ -22,6 +20,42 @@ public class ProductConfigurationProperties {
     private Cursor cursor;
     private IngestionMode ingestionMode = new IngestionMode();
 
+    public boolean isCompletedEventEnabled() {
+        return Boolean.TRUE.equals(events.getEnableCompleted());
+    }
+
+    public boolean isFailedEventEnabled() {
+        return Boolean.TRUE.equals(events.getEnableFailed());
+    }
+
+    public boolean isTransactionChainEnabled() {
+        return Boolean.TRUE.equals(chains.getTransactionComposition().getEnabled());
+    }
+
+    public boolean isTransactionChainEnabled(RequestConfig requestConfig) {
+        return requestConfig == null || requestConfig.isTransactionChainEnabled().isEmpty()
+                ? Boolean.TRUE.equals(chains.getTransactionComposition().getEnabled())
+                : requestConfig.isTransactionChainEnabled().orElse(false);
+    }
+
+    public boolean isTransactionChainAsync(RequestConfig requestConfig) {
+        return requestConfig == null || requestConfig.isTransactionChainAsync().isEmpty()
+                ? Boolean.TRUE.equals(chains.getTransactionComposition().getAsync())
+                : requestConfig.isTransactionChainAsync().orElse(false);
+    }
+
+    public boolean isTransactionChainAsync() {
+        return Boolean.TRUE.equals(chains.getTransactionComposition().getAsync());
+    }
+
+    public boolean isPaymentOrderChainEnabled() {
+        return Boolean.TRUE.equals(chains.getPaymentOrderComposition().getEnabled());
+    }
+
+    public boolean isPaymentOrderChainAsync() {
+        return Boolean.TRUE.equals(chains.getPaymentOrderComposition().getAsync());
+    }
+
     @Data
     @NoArgsConstructor
     public static class Events {
@@ -37,7 +71,9 @@ public class ProductConfigurationProperties {
     }
 
     @Data
+    @Builder
     @NoArgsConstructor
+    @AllArgsConstructor
     public static class Chains {
         private TransactionComposition transactionComposition;
         private PaymentOrderComposition paymentOrderComposition;
@@ -52,45 +88,28 @@ public class ProductConfigurationProperties {
     }
 
     @Data
+    @SuperBuilder
+    @NoArgsConstructor
+    @AllArgsConstructor
     public static abstract class BaseComposition {
         private Boolean enabled = Boolean.FALSE;
         private String baseUrl = "http://localhost:9003/";
         private Boolean async = Boolean.FALSE;
     }
 
-    @NoArgsConstructor
     @Data
+    @NoArgsConstructor
+    @AllArgsConstructor
+    @SuperBuilder
     public static class TransactionComposition extends BaseComposition {
         private List<String> excludeProductTypeExternalIds = new ArrayList<>();
     }
 
-    @NoArgsConstructor
     @Data
+    @NoArgsConstructor
+    @AllArgsConstructor
+    @SuperBuilder
     public static class PaymentOrderComposition extends BaseComposition {
         private List<String> excludeProductTypeExternalIds = new ArrayList<>();
-    }
-
-    public boolean isCompletedEventEnabled() {
-        return Boolean.TRUE.equals(events.getEnableCompleted());
-    }
-
-    public boolean isFailedEventEnabled() {
-        return Boolean.TRUE.equals(events.getEnableFailed());
-    }
-
-    public boolean isTransactionChainEnabled() {
-        return Boolean.TRUE.equals(chains.getTransactionComposition().getEnabled());
-    }
-
-    public boolean isTransactionChainAsync() {
-        return Boolean.TRUE.equals(chains.getTransactionComposition().getAsync());
-    }
-
-    public boolean isPaymentOrderChainEnabled() {
-        return Boolean.TRUE.equals(chains.getPaymentOrderComposition().getEnabled());
-    }
-
-    public boolean isPaymentOrderChainAsync() {
-        return Boolean.TRUE.equals(chains.getPaymentOrderComposition().getAsync());
     }
 }