+++ stream-compositions/services/product-composition-service/src/main/java/com/backbase/stream/compositions/product/core/model/ArrangementIngestPullRequest.java
@@ -1,0 +1,13 @@
+package com.backbase.stream.compositions.product.core.model;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class ArrangementIngestPullRequest {
+    private String arrangementId;
+    private String externalArrangementId;
+    private String source;
+    private RequestConfig config;
+}