+++ stream-compositions/services/product-composition-service/src/main/java/com/backbase/stream/compositions/product/core/service/impl/ProductIngestionServiceImpl.java
@@ -34,9 +34,7 @@ public class ProductIngestionServiceImpl implements ProductIngestionService {
     private final BatchProductIngestionSaga batchProductIngestionSaga;
     private final ProductIntegrationService productIntegrationService;
     private final ProductConfigurationProperties config;
-
     private final Validator validator;
-
     private final ProductPostIngestionService productPostIngestionService;
 
     /**