+++ stream-compositions/services/product-composition-service/src/test/java/com/backbase/stream/compositions/product/http/ProductControllerIT.java
@@ -1,13 +1,10 @@
 package com.backbase.stream.compositions.product.http;
 
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.when;
-import static org.mockserver.integration.ClientAndServer.startClientAndServer;
-import static org.mockserver.model.HttpRequest.request;
-import static org.mockserver.model.HttpResponse.response;
-
+import com.backbase.dbs.arrangement.api.service.v2.model.AccountArrangementItemPut;
 import com.backbase.stream.compositions.paymentorder.client.model.PaymentOrderIngestionResponse;
 import com.backbase.stream.compositions.paymentorder.client.model.PaymentOrderPostResponse;
+import com.backbase.stream.compositions.product.api.model.ArrangementPullIngestionRequest;
+import com.backbase.stream.compositions.product.api.model.ArrangementPushIngestionRequest;
 import com.backbase.stream.compositions.product.api.model.ProductPullIngestionRequest;
 import com.backbase.stream.compositions.product.api.model.ProductPushIngestionRequest;
 import com.backbase.stream.compositions.transaction.client.model.TransactionIngestionResponse;
@@ -16,17 +13,13 @@ import com.backbase.stream.legalentity.model.BatchProductGroup;
 import com.backbase.stream.legalentity.model.ProductGroup;
 import com.backbase.stream.legalentity.model.ServiceAgreement;
 import com.backbase.stream.product.BatchProductIngestionSaga;
+import com.backbase.stream.product.service.ArrangementService;
 import com.backbase.stream.product.task.BatchProductGroupTask;
 import com.backbase.stream.product.task.ProductGroupTask;
 import com.backbase.streams.compositions.test.IntegrationTest;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import java.io.IOException;
-import java.net.URI;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.TimeUnit;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.activemq.broker.BrokerService;
 import org.junit.jupiter.api.AfterEach;
@@ -47,6 +40,18 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.reactive.server.WebTestClient;
 import reactor.core.publisher.Mono;
 
+import java.io.IOException;
+import java.net.URI;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.TimeUnit;
+
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.when;
+import static org.mockserver.integration.ClientAndServer.startClientAndServer;
+import static org.mockserver.model.HttpRequest.request;
+import static org.mockserver.model.HttpResponse.response;
+
 @DirtiesContext
 @SpringBootTest
 @AutoConfigureWebTestClient
@@ -57,23 +62,22 @@ class ProductControllerIT extends IntegrationTest {
     private static final int INTEGRATION_SERVICE_PORT = 18000;
     private static final int TRANSACTION_SERVICE_PORT = 12000;
     private static final int PAYMENT_ORDER_SERVICE_PORT = 13000;
+    private static BrokerService broker;
+    @MockBean
+    @Qualifier("batchProductIngestionSaga")
+    BatchProductIngestionSaga batchProductIngestionSaga;
+    @MockBean
+    ArrangementService arrangementService;
+    @Autowired
+    ProductController productController;
+    @Autowired
+    ObjectMapper objectMapper;
     private ClientAndServer integrationServer;
     private ClientAndServer transactionServer;
     private ClientAndServer paymentOrderServer;
     private MockServerClient integrationServerClient;
     private MockServerClient transactionServerClient;
     private MockServerClient paymentOrderServerClient;
-    private static BrokerService broker;
-
-    @MockBean
-    @Qualifier("batchProductIngestionSaga")
-    BatchProductIngestionSaga batchProductIngestionSaga;
-
-    @Autowired
-    ProductController productController;
-
-    @Autowired
-    ObjectMapper objectMapper;
 
     @BeforeAll
     static void initActiveMqBroker() throws Exception {
@@ -97,6 +101,17 @@ class ProductControllerIT extends IntegrationTest {
                                 .withStatusCode(200)
                                 .withContentType(MediaType.APPLICATION_JSON)
                                 .withBody(readContentFromClasspath("integration-data/response.json"))
+                );
+
+        integrationServerClient.when(
+                        request()
+                                .withMethod("PUT")
+                                .withPath("/service-api/v2/arrangement"))
+                .respond(
+                        response()
+                                .withStatusCode(200)
+                                .withContentType(MediaType.APPLICATION_JSON)
+                                .withBody(readContentFromClasspath("integration-data/arrangement-response.json"))
                 );
     }
 
@@ -154,7 +169,7 @@ class ProductControllerIT extends IntegrationTest {
     void pullIngestProduct_Success() throws Exception {
         ObjectMapper mapper = new ObjectMapper();
         JsonNode node = mapper.readTree(readContentFromClasspath("integration-data/response.json"))
-            .get("productGroups").get(0);
+                .get("productGroups").get(0);
         ProductGroup productGroup = mapper.treeToValue(node, ProductGroup.class);
         productGroup.setServiceAgreement(new ServiceAgreement().internalId("sa_internalId"));
 
@@ -184,7 +199,7 @@ class ProductControllerIT extends IntegrationTest {
 
         webTestClient.post().uri(uri)
                 .contentType(org.springframework.http.MediaType.APPLICATION_JSON)
-                .body(Mono.just(pullIngestionRequest), ProductPullIngestionRequest.class). exchange()
+                .body(Mono.just(pullIngestionRequest), ProductPullIngestionRequest.class).exchange()
                 .expectStatus().isCreated();
     }
 
@@ -199,5 +214,70 @@ class ProductControllerIT extends IntegrationTest {
                 .contentType(org.springframework.http.MediaType.APPLICATION_JSON)
                 .body(Mono.just(pushIngestionRequest), ProductPushIngestionRequest.class).exchange()
                 .expectStatus().is5xxServerError();
+    }
+
+    @Test
+    void pullIngestArrangement_Success() throws Exception {
+        when(arrangementService.updateArrangement(any()))
+                .thenReturn(Mono.just(new AccountArrangementItemPut()));
+
+        ArrangementPullIngestionRequest pullIngestionRequest =
+                new ArrangementPullIngestionRequest()
+                        .withInternalArrangementId("arrangementId")
+                        .withExternalArrangementId("externalArrangementId")
+                        .withAdditions(Map.of());
+
+        URI uri = URI.create("/service-api/v2/ingest/arrangement/pull");
+        WebTestClient webTestClient = WebTestClient.bindToController(productController).build();
+
+        webTestClient.put().uri(uri)
+                .contentType(org.springframework.http.MediaType.APPLICATION_JSON)
+                .body(Mono.just(pullIngestionRequest), ProductPullIngestionRequest.class).exchange()
+                .expectStatus().isOk();
+    }
+
+    @Test
+    void pullIngestArrangement_Fail() throws Exception {
+        when(arrangementService.updateArrangement(any()))
+                .thenThrow(new RuntimeException());
+
+        ArrangementPullIngestionRequest pullIngestionRequest =
+                new ArrangementPullIngestionRequest()
+                        .withInternalArrangementId("arrangementId")
+                        .withExternalArrangementId("externalArrangementId")
+                        .withAdditions(Map.of());
+
+        URI uri = URI.create("/service-api/v2/ingest/arrangement/pull");
+        WebTestClient webTestClient = WebTestClient.bindToController(productController).build();
+
+        webTestClient.put().uri(uri)
+                .contentType(org.springframework.http.MediaType.APPLICATION_JSON)
+                .body(Mono.just(pullIngestionRequest), ProductPullIngestionRequest.class).exchange()
+                .expectStatus().is5xxServerError();
+    }
+
+    @Test
+    void pushIngestArrangement_Success() throws IOException {
+        ObjectMapper mapper = new ObjectMapper();
+        JsonNode node = mapper.readTree(readContentFromClasspath("integration-data/arrangement-response.json"))
+                .get("arrangement");
+        com.backbase.stream.compositions.product.api.model.AccountArrangementItemPut arrangementItemPut =
+                mapper.treeToValue(node, com.backbase.stream.compositions.product.api.model.AccountArrangementItemPut.class);
+
+        when(arrangementService.updateArrangement(any()))
+                .thenReturn(Mono.just(new AccountArrangementItemPut()));
+
+        ArrangementPushIngestionRequest pushIngestionRequest =
+                new ArrangementPushIngestionRequest()
+                        .withInternalArrangementId("arrangementId")
+                        .withArrangement(arrangementItemPut);
+
+        URI uri = URI.create("/service-api/v2/ingest/arrangement/push");
+        WebTestClient webTestClient = WebTestClient.bindToController(productController).build();
+
+        webTestClient.put().uri(uri)
+                .contentType(org.springframework.http.MediaType.APPLICATION_JSON)
+                .body(Mono.just(pushIngestionRequest), ArrangementPushIngestionRequest.class).exchange()
+                .expectStatus().isOk();
     }
 }